Stop GCloud initilization early if service account key var not found

+++ server/services/gcloud/tasks.go
@@ -22,11 +22,15 @@ import (
 var TasksClient *cloudtasks.Client
 
 func InitTasks() func() {
+	credsFile := os.Getenv("SERVICE_ACCOUNT_KEY_PATH")
+	if credsFile == "" || credsFile == "?" {
+		logger.StdErr.Println("SERVICE_ACCOUNT_KEY_PATH not set, skipping Cloud Tasks initialization")
+		return func() {}
+	}
+
 	ctx := context.Background()
 
 	var err error
-	credsFile := os.Getenv("SERVICE_ACCOUNT_KEY_PATH")
-
 	TasksClient, err = cloudtasks.NewClient(ctx, option.WithCredentialsFile(credsFile))
 	if err != nil {
 		logger.StdErr.Panicln(err)
@@ -39,6 +43,11 @@ func InitTasks() func() {
 }
 
 func CreateEmailTask(email string, ownerName string, eventName string, eventId string) []string {
+	if TasksClient == nil {
+		logger.StdErr.Println("WARNING: Cloud Tasks is disabled, skipping CreateEmailTask")
+		return []string{}
+	}
+
 	// Get listmonk url env vars
 	listmonkUrl := os.Getenv("LISTMONK_URL")
 	listmonkUsername := os.Getenv("LISTMONK_USERNAME")
@@ -127,6 +136,11 @@ func CreateEmailTask(email string, ownerName string, eventName string, eventId s
 }
 
 func DeleteEmailTask(taskId string) {
+	if TasksClient == nil {
+		logger.StdErr.Println("WARNING: Cloud Tasks is disabled, skipping DeleteEmailTask")
+		return
+	}
+
 	err := TasksClient.DeleteTask(context.Background(), &cloudtaskspb.DeleteTaskRequest{
 		Name: taskId,
 	})
